# PERF: Improve performance for df.duplicated with one column subset